requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
discord.py==2.6.4
distro==1.9.0
duckduckgo-search==7.1.0
easy-pil==0.4.0
frozenlist==1.8.0
googlesearch-python==1.3.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
multidict==6.7.0
pillow==10.4.0
primp==0.15.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wikipedia==1.4.0
yarl==1.22.0
